+++ centreon/packages/js-config/cypress/e2e/configuration.ts
@@ -4,53 +4,62 @@
 import { execSync } from 'child_process';
 
 import { defineConfig } from 'cypress';
+import createBundler from '@bahmutov/cypress-esbuild-preprocessor';
 import Docker from 'dockerode';
 import { addCucumberPreprocessorPlugin } from '@badeball/cypress-cucumber-preprocessor';
 import webpackPreprocessor from '@cypress/webpack-preprocessor';
+import createEsbuildPlugin from '@badeball/cypress-cucumber-preprocessor/esbuild';
+import cypressOnFix from 'cypress-on-fix';
 
 const docker = new Docker();
 
-const getWebpackOptions = (config): object => {
-  return {
-    module: {
-      rules: [
-        {
-          exclude: [/node_modules/],
-          test: /\.ts?$/,
-          use: [
-            {
-              loader: 'swc-loader'
-            }
-          ]
-        },
-        {
-          test: /\.feature$/,
-          use: [
-            {
-              loader: '@badeball/cypress-cucumber-preprocessor/webpack',
-              options: config
-            }
-          ]
-        }
-      ]
-    },
-    resolve: {
-      extensions: ['.ts', '.js']
-    }
-  };
-};
-
-const setupNodeEvents = async (cypressOn, config): Promise<void> => {
-  const on = require('cypress-on-fix')(cypressOn);
+export const setupNodeEvents = async (
+  cypressOn: Cypress.PluginEvents,
+  config: Cypress.PluginConfigOptions
+): Promise<Cypress.PluginConfigOptions> => {
+  const on = cypressOnFix(cypressOn);
 
   await addCucumberPreprocessorPlugin(on, config);
 
-  const webpackOptions = await getWebpackOptions(config);
-  const options = {
-    webpackOptions
-  };
+  on(
+    'file:preprocessor',
+    createBundler({
+      plugins: [createEsbuildPlugin(config)]
+    })
+  );
 
-  on('file:preprocessor', webpackPreprocessor(options));
+  // on(
+  //   'file:preprocessor',
+  //   webpackPreprocessor({
+  //     webpackOptions: {
+  //       module: {
+  //         rules: [
+  //           {
+  //             exclude: [/node_modules/],
+  //             test: /\.ts?$/,
+  //             use: [
+  //               {
+  //                 loader: 'swc-loader'
+  //               }
+  //             ]
+  //           },
+  //           {
+  //             test: /\.feature$/,
+  //             use: [
+  //               {
+  //                 loader: '@badeball/cypress-cucumber-preprocessor/webpack',
+  //                 options: config
+  //               }
+  //             ]
+  //           }
+  //         ]
+  //       },
+  //       resolve: {
+  //         extensions: ['.ts', '.js']
+  //       }
+  //     }
+  //   })
+  // );
 
   on('before:browser:launch', (browser = {}, launchOptions) => {
     if ((browser as { name }).name === 'chrome') {
@@ -84,11 +93,17 @@ const setupNodeEvents = async (cypressOn, config): Promise<void> => {
       name,
       portBindings = []
     }: StartContainerProps) => {
+      console.log(`Starting container ${image}`);
+
       const webContainers = await docker.listContainers({
         all: true,
         filters: { name: [name] }
       });
       if (webContainers.length) {
+        console.log(
+          `Container ${image} already started : baseUrl has probably changed and reloaded the test`
+        );
+
         return webContainers[0];
       }
 
@@ -122,12 +137,19 @@ const setupNodeEvents = async (cypressOn, config): Promise<void> => {
 
       await container.start();
 
+      console.log(`Container ${image} started`);
+
       return container;
     },
     stopContainer: async ({ name }: StopContainerProps) => {
       const container = await docker.getContainer(name);
+
+      console.log(`Stopping container ${name}`);
+
       await container.kill();
       await container.remove();
+
+      console.log(`Container ${name} killed and removed`);
 
       return null;
     }
@@ -164,6 +186,9 @@ export default ({
     defaultCommandTimeout: 6000,
     e2e: {
       excludeSpecPattern: ['*.js', '*.ts', '*.md'],
+      reporter: require.resolve(
+        '@badeball/cypress-cucumber-preprocessor/pretty-reporter'
+      ),
       setupNodeEvents,
       specPattern
     },
@@ -175,6 +200,8 @@ export default ({
       dockerName: dockerName || 'centreon-dev'
     },
     execTimeout: 60000,
+    reporter: require.resolve('cypress-multi-reporters'),
+    /*
     reporter: 'mochawesome',
     reporterOptions: {
       html: false,
@@ -183,6 +210,7 @@ export default ({
       reportDir: `${resultsFolder}/reports`,
       reportFilename: '[name]-report.json'
     },
+    */
     requestTimeout: 10000,
     retries: 0,
     screenshotsFolder: `${resultsFolder}/screenshots`,
