+++ centreon/tests/e2e/cypress/support/e2e.ts
@@ -1,6 +1,15 @@
 import 'cypress-wait-until';
 import './commands';
 
+before(() => {
+  Cypress.config('baseUrl', 'http://0.0.0.0:4000');
+
+  cy.intercept('/waiting-page', {
+    headers: { 'content-type': 'text/html' },
+    statusCode: 200
+  }).visit('/waiting-page');
+});
+
 Cypress.on('uncaught:exception', (err) => {
   if (
     err.message.includes('Request failed with status code 401') ||
